requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
requests
pytesseract
Pillow
firebase-admin
psycopg2-binary
